fix: 휴가 캘린더 API 업데이트 (250731)
fix: 휴가 캘린더 응답 및 QueryDSL projection에서 daysCount 타입 변경 (int → BigDecimal)

+++ src/main/java/com/melly/vacationmanager/domain/vacation/request/dto/response/VacationCalendarResponse.java
@@ -2,6 +2,7 @@ package com.melly.vacationmanager.domain.vacation.request.dto.response;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -14,6 +15,6 @@ public class VacationCalendarResponse {
     private String typeName;
     private LocalDate startDate;
     private LocalDate endDate;
-    private int daysCount;
+    private BigDecimal daysCount;
     private String status;
 }
